Add assert and logging to clean and configure scripts

+++ fixes/clean.sh
@@ -1,3 +1,9 @@
+if [ ! -f 'bug-info/original-name' ]
+then
+    echo FAIL
+    exit 1
+fi
+
 rm -rf fixed-program.txt \
     *.cache \
     *~ \
@@ -19,8 +25,21 @@ mv fix.lines bug-info
 mv fault.lines bug-info
 mv bugged-program.txt bug-info
 
+# Fix occurrences of manybugs VM paths sprayed throughout the benchmark program
+manybugs_dir="/root/mountpoint-genprog/genprog-many-bugs/$(cat bug-info/original-name)"
 cp test.sh.back test.sh
-if [ -d test.sh ]
+if [ -f test.sh ]
 then
-    sed -i.back "s@/root/mountpoint-genprog/genprog-many-bugs/$(cat bug-info/original-name)/@../@g" test.sh
+    sed -i.back "s@$manybugs_dir@../@g" test.sh
 fi
+
+for b in `grep --include='*.back' -Rl $manybugs_dir`
+do
+    echo "Restoring $b"
+    cp $b ${b%%.back}
+done
+for f in `grep --exclude='*.back' -Rl $manybugs_dir`
+do
+    echo "Replacing $f: $manybugs_dir -> $PWD"
+    sed -i.back "s@$manybugs_dir@$PWD@g" $f
+done
